Revise curl mbedtls fix
This is the official fix, see https://github.com/curl/curl/pull/15846. The main difference is that we'll have to document our sockets as being picky about what is passed to them when retrying writes: the exact same must be passed as the previous time, potentially followed by new data.