Make zlib an indirect dependency
The game does nothing with the zlib API, yet it used to declare it as a dependency and get access to headers and libraries that way. This commit makes it possible for the game to stop doing that. TODO: Somehow move the conditional declaration of the project-wide ZLIB_WINAPI macro over to tpt-libs too.